Use resource builders to genericise access to Kubernetes
